==> 1. GenerateTrafficWeatherEvent2Cities.ipynb <==
### 1 -Importons les différentes librairies 

#!pip install tensorflow
import pandas as pd
import numpy as np
from datetime import datetime
import time
import warnings
warnings.filterwarnings("ignore")

#import tarfile
#file = tarfile.open("data/traffic_weather_ug16_june19.tar.gz", mode="r:gz")
#file.extractall()
#file.close()

### 2 - Importons le Dataset "TrafficWeatherEvent_Aug16_June19_Publish.csv"

tw_event = pd.read_csv("data/TrafficWeatherEvent_Aug16_June19_Publish.csv")

tw_event.tail()

### Récupérons les coordonnées géographiques de chaque ville dans un dictionnaire afin d'afficher la carte

#geo_dic = {}
#for i in range(len(tw_event)):
#    cities = tw_event.City.iloc[i]
#    lat = tw_event.LocationLat.iloc[i]
#    lng = tw_event.LocationLng.iloc[i]
#    geo_dic["{}".format(cities)] = [lat, lng]

#### Choix max et min des latitudes et longitudes de chaque ville

#LosAngeles = df_cities[df_cities.City == "Los Angeles"]
#latL = LosAngeles.LocationLat
#lngL = LosAngeles.LocationLng
#np.min(latL), np.max(latL), np.min(lngL), np.max(lngL)

### 3 - 1 - Choix des villes sur le dataset

cities = {'LosAngeles': [33.700615, 34.353627, -118.683511, -118.074559], 
           'Houston': [29.497907,30.129003,-95.797178,-94.988191],
           'Austin': [30.079327, 30.596764,-97.968881,-97.504838],
           'Dallas': [32.559567,33.083278,-97.036586,-96.428928],
           'Charlotte': [34.970168,35.423667,-81.060925,-80.622687],
           'Atlanta': [33.612410,33.916999,-84.575600,-84.231911]}

time_zones = {'Houston':'US/Central', 'Charlotte':'US/Eastern', 'Dallas':'US/Central',
              'Atlanta':'US/Eastern', 'Austin':'US/Central', 'LosAngeles':'US/Pacific'}

# time interval to sample data for 
start = datetime(2018, 6, 1)
finish   = datetime(2018, 9, 2)

begin = datetime.strptime('2018-06-01 00:00:00', '%Y-%m-%d %H:%M:%S')
end   = datetime.strptime('2018-08-31 23:59:59', '%Y-%m-%d %H:%M:%S')

### 3 - 2 - Le Dataset du Trafic Event pour toutes les Villes Choisies

tw_event['StartTime(UTC)'] = tw_event['StartTime(UTC)'].astype('datetime64[ns]', errors = 'ignore')
tw_event['EndTime(UTC)'] = tw_event['EndTime(UTC)'].astype('datetime64[ns]', errors = 'ignore')

for c in cities:
    crds = cities[c]
    subset_all = tw_event[(tw_event['Source'] == 'T') & (tw_event['StartTime(UTC)'] >= start) & (tw_event['StartTime(UTC)'] < end) & 
                    (tw_event['LocationLat']>crds[0]) & (tw_event['LocationLat']<crds[1]) & (tw_event['LocationLng']>crds[2]) & 
                    (tw_event['LocationLng']<crds[3])& (tw_event["AirportCode"] != "KMWS")]
    
    subset_accidents = tw_event[(tw_event['Type']=='Accident') &  (tw_event['Source'] == 'T') &(tw_event['StartTime(UTC)'] >= start) & (tw_event['StartTime(UTC)'] < end) 
                          & (tw_event['LocationLat']>crds[0]) & (tw_event['LocationLat']<crds[1]) & (tw_event['LocationLng']>crds[2]) 
                          & (tw_event['LocationLng']<crds[3])]
    
    print("Pour {} nous avons {} incidents, dont {} accidents! un taux de {:.2f}".format(c, len(subset_all), len(subset_accidents), 
                                                                               len(subset_accidents)*1.0/len(subset_all)))
    
    subset_all.to_csv('data/temporary/TW_EVENT_{}_20180601_20180609.csv'.format(c), index=False)

### !!! Fin de generation de TW_Event_Cities, pour la suite, voir le fichier GenerateFeaturesVectors.ipynb  !!!
