==> 4. Creer des Datasets Sample_Geohash_City_TW_Event_Poi.ipynb <==
### 1- Importation des librairies 

import pandas as pd
import numpy as np


### 2- Choix des villes et intervalles de Coordonnées Géographiques

cities = {'LosAngeles': [33.700615, 34.353627, -118.683511, -118.074559], 
           'Houston': [29.497907,30.129003,-95.797178,-94.988191],
           'Austin': [30.079327, 30.596764,-97.968881,-97.504838],
           'Dallas': [32.559567,33.083278,-97.036586,-96.428928],
           'Charlotte': [34.970168,35.423667,-81.060925,-80.622687],
           'Atlanta': [33.612410,33.916999,-84.575600,-84.231911]}

### 3- Importation du dataset geohash_to_poi

geohash_map = pd.read_csv("data/geohash_to_poi_vec.csv")
geohash_map.tail()

### 4- Jointure et concatenation des datasets 

for c in cities :
    df = pd.read_csv("data/vectors/{}_geo2vec_201861-2018831.csv".format(c))
    df = pd.merge(df, geohash_map)
    df= df.loc[:,"Geohash":]
    df.to_csv("data/TW_Event_Poi_vectors/Geohash_{}_TW_Event_Poi.csv".format(c), index=False)

for c in cities :
    df = pd.read_csv("data/TW_Event_Poi_vectors/Geohash_{}_TW_Event_Poi.csv".format(c))
    df_0 = df[df["T-Accident"] == 0].sample(13000)
    df_1 = df[df["T-Accident"] != 0]
    df_sample = pd.concat([df_0, df_1])
    df_sample.to_csv("data/Sample_TW_Event_Poi_vectors/Sample_Geohash_{}_TW_Event_Poi.csv".format(c), index=False)

#### 5- Importation des  Datasets

Atlanta = pd.read_csv("data/Sample_TW_Event_Poi_vectors/Sample_Geohash_Atlanta_TW_Event_Poi.csv")
#Atlanta = Atlanta.loc[:,"Geohash":]
Charlotte = pd.read_csv("data/Sample_TW_Event_Poi_vectors/Sample_Geohash_Charlotte_TW_Event_Poi.csv")
#Charlotte = Charlotte.loc[:,"Geohash":]
Dallas = pd.read_csv("data/Sample_TW_Event_Poi_vectors/Sample_Geohash_Dallas_TW_Event_Poi.csv")
#Dallas = Dallas.loc[:,"Geohash":]
Austin = pd.read_csv("data/Sample_TW_Event_Poi_vectors/Sample_Geohash_Austin_TW_Event_Poi.csv")
#Austin = Austin.loc[:,"Geohash":]
Houston = pd.read_csv("data/Sample_TW_Event_Poi_vectors/Sample_Geohash_Houston_TW_Event_Poi.csv")
#Houston = Houston.loc[:,"Geohash":]
Losangeles = pd.read_csv("data/Sample_TW_Event_Poi_vectors/Sample_Geohash_LosAngeles_TW_Event_Poi.csv")
#Losangeles = Losangeles.loc[:,"Geohash":]

#### 6- Nombre total d'Incidents et le taux d'Accidents par Ville

for c in cities :
    if c == "Atlanta":
        incident = len(Atlanta)
        accident = len(Atlanta[Atlanta["T-Accident"] != 0])
        print("Pour la ville de {}, nous avons enrégistré {} incidents dont {} accidents, un taux de {}".format(c, incident, accident, accident/incident))
    elif c == "Charlotte":
        incident = len(Charlotte)
        accident = len(Charlotte[Charlotte["T-Accident"] != 0])
        print("Pour la ville de {}, nous avons enrégistré {} incidents dont {} accidents, un taux de {}".format(c, incident, accident, accident/incident))
    elif c == "Dallas":
        incident = len(Dallas)
        accident = len(Dallas[Dallas["T-Accident"] != 0])
        print("Pour la ville de {}, nous avons enrégistré {} incidents dont {} accidents, un taux de {}".format(c, incident, accident, accident/incident))
    elif c == "Austin":
        incident = len(Austin)
        accident = len(Austin[Austin["T-Accident"] != 0])
        print("Pour la ville de {}, nous avons enrégistré {} incidents dont {} accidents, un taux de {}".format(c, incident, accident, accident/incident))
    elif c == "Houston":
        incident = len(Houston)
        accident = len(Houston[Houston["T-Accident"] != 0])
        print("Pour la ville de {}, nous avons enrégistré {} incidents dont {} accidents, un taux de {}".format(c, incident, accident, accident/incident))
    else :
        incident = len(Losangeles)
        accident = len(Losangeles[Losangeles["T-Accident"] != 0])
        print("Pour la ville de {}, nous avons enrégistré {} incidents dont {} accidents, un taux de {}".format(c, incident, accident, accident/incident))
