==> 5. Prediction des Accidents de Trafic Avec les datasets de Sample_WT_Event_Poi_vectors.ipynb <==
### 1- Importer des librairies

!pip install tensorflow
!pip install keras

import pandas as pd
import numpy as np

import matplotlib
import matplotlib.pyplot as plt
import math

import tensorflow as tf
import keras

from keras import Sequential
from keras.layers import Dense

from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import classification_report, confusion_matrix
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.model_selection import GridSearchCV

import warnings
warnings.filterwarnings("ignore")

#### 2- Choix des Villes et les Intervalles de Coordonnées Géographiques

cities = {'LosAngeles': [33.700615, 34.353627, -118.683511, -118.074559], 
           'Houston': [29.497907,30.129003,-95.797178,-94.988191],
           'Austin': [30.079327, 30.596764,-97.968881,-97.504838],
           'Dallas': [32.559567,33.083278,-97.036586,-96.428928],
           'Charlotte': [34.970168,35.423667,-81.060925,-80.622687],
           'Atlanta': [33.612410,33.916999,-84.575600,-84.231911]}

#### 3- Pre-processing de tous les datasets

geohash_map = pd.read_csv("data/geohash_to_poi_vec.csv")
geo_dict = dict(zip(geohash_map.Geohash.unique(), range(len(geohash_map.Geohash.unique()))))

def clean_data(filepath, cityname):
    df = pd.read_csv(filepath)
    print ("Le taux de zero accident dans la ville de {} est égal à :".format(cityname),float(df[df['T-Accident']==0].shape[0])/df.shape[0])
    def fun_hash(geohash):
        return geo_dict[geohash]
    df['geohash_code'] = df.apply(lambda row: fun_hash(row['Geohash']), axis=1)
    def week_day(DOW):
        if DOW < 5:
            return 1
        else:
            return 0
    def shift(group):
        df_list=[]
        for idx,df in group:
            df['predicted_accident'] = df['T-Accident'].shift(-1)
            df.drop(df.tail(1).index,inplace=True)
            df_list.append(df)
        return pd.concat(df_list)

    def time_interval(HOD):
        if HOD >=6 and HOD <10:
            return 0
        if HOD >= 10 and HOD<15:
            return 1
        if HOD >=15 and HOD< 18:
            return 2;
        if HOD >=18 and HOD< 22:
            return 3
        else:
            return 4; 
    def make_binary(d):
        if d > 0:
            return 1
        else:
            return 0    
    df['DOW'] = df.apply(lambda row: week_day(row['DOW']), axis=1)   
    df['HOD'] = df.apply(lambda row: time_interval(row['HOD']), axis=1) 
    df['T-Accident'] = df.apply(lambda row: make_binary(row['T-Accident']), axis=1) 
    group = df.groupby('Geohash')
    df = shift(group)
    return df.to_csv("data/Clean_TW_Data/{}_Clean_TW_Data.csv".format(cityname), index=False)

LosAngeles = pd.read_csv("data/Clean_TW_Data/LosAngeles_Clean_TW_Data.csv")
Houston = pd.read_csv("data/Clean_TW_Data/Houston_Clean_TW_Data.csv")
Austin = pd.read_csv("data/Clean_TW_Data/Austin_Clean_TW_Data.csv")
Dallas = pd.read_csv("data/Clean_TW_Data/Dallas_Clean_TW_Data.csv")
Charlotte = pd.read_csv("data/Clean_TW_Data/Charlotte_Clean_TW_Data.csv")
Atlanta = pd.read_csv("data/Clean_TW_Data/Atlanta_Clean_TW_Data.csv")

LosAngeles.tail()

for city in cities:
    clean_data("data/Sample_TW_Event_Poi_vectors/Sample_Geohash_{}_TW_Event_Poi.csv".format(city), city)

### 3-1- Concaténation des datasets

for city in cities :
    df = pd.read_csv("data/Clean_TW_Data/{}_Clean_TW_Data.csv".format(city))
    df = df.drop(["Geohash", "T-Accident","T-BrokenVehicle","T-Congestion","T-Construction","T-Event", "T-FlowIncident", "T-RoadBlocked", "T-Other"], axis = 1)
    df.to_csv("data/clean_twpoi_data/{}.csv".format(city), index = False)

df = pd.read_csv("data/clean_twpoi_data/Atlanta.csv")

df.tail()

dataset = pd.concat([LosAngeles, Houston, Austin, Dallas, Charlotte, Atlanta])

dataset.to_csv("data/clean_twpoi_data/clean_twpoi_data.csv", index=False)

df = pd.read_csv("data/clean_twpoi_data/clean_twpoi_data.csv")

print ("Le taux de zero accident  est égal à :" ,float(df[df['predicted_accident']==0].shape[0])/df.shape[0])

df = pd.read_csv("data/clean_twpoi_data/TrafficWeatherEvent_June18_Aug18_Publish.csv")
df.tail()

### 3-2- Importation du dataset

tw = pd.read_csv("data/clean_twpoi_data/TrafficWeatherEvent_June18_Aug18_Publish.csv")
tw.tail()

### 4- Models de prédictions

#### 4-1- Modèle de Regression Logistic

def logistic_regression(filepath, cityname):
    df = pd.read_csv(filepath)
    X = df.loc[:, df.columns != "predicted_accident"]
    y = df.loc[:, df.columns == "predicted_accident"]
    
    # Split in train/test
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, stratify = y, random_state = 0)
    
    

    # Standardisize
    sc_x = StandardScaler()
    X_train = sc_x.fit_transform(X_train)
    X_test = sc_x.transform(X_test)
    
    # Modele de regression Logistique
    print("{},  logistic regression  ... ".format(cityname))
    classifier = LogisticRegression( )
    classifier.fit(X_train, y_train)
    print()
    print("{}, score de performance du modèle (train) : {}  ".format(cityname, classifier.score(X_train, y_train)))
    print("{}, score de performance du modèle (test) : {}  ".format(cityname, classifier.score(X_test, y_test)))
    print()
    predictions = classifier.predict(X_test)
    print("Confusion Matrix:")
    print(confusion_matrix(y_test, predictions))
    print()
    print("Classification Report")
    print(classification_report(y_test, predictions))
    print("....Done")
    print()

c = {"TrafficWeatherEvent_June18_Aug18_Publish": 1}

for city in c :
    logistic_regression("data/clean_twpoi_data/{}.csv".format(city), city)

for city in cities :
    logistic_regression("data/clean_twpoi_data/{}.csv".format(city), city)  

#### 4-2- Modèle de Gradient Boosting Classifier

def gradient_boosting(filepath, cityname):
    df = pd.read_csv(filepath)
    X = df.loc[:, df.columns != "predicted_accident"]
    y = df.loc[:, df.columns == "predicted_accident"]
    
    # Split in train/test
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.3, stratify = y, random_state = 42)
    
    

    # Standardisize
    sc_x = StandardScaler()
    X_train = sc_x.fit_transform(X_train)
    X_test = sc_x.transform(X_test)
    
    # Gradient Boosting Classifier
    print("{},  Gradient Boosting Classifier  ... ".format(cityname))
   
    gb_clf2 = GradientBoostingClassifier(n_estimators = 90, learning_rate=0.95, 
                                        random_state=0)
    gb_clf2.fit(X_train, y_train)
    print()
    print("Accuracy score (training): {0:.3f}".format(gb_clf2.score(X_train, y_train)))
    print("Accuracy score (test): {0:.3f}".format(gb_clf2.score(X_test, y_test)))
    print()
    
    predictions = gb_clf2.predict(X_test)

    print()
    print("Confusion Matrix:")
    print(confusion_matrix(y_test, predictions))
    
    print()
    print("Classification Report")
    print(classification_report(y_test, predictions))
    print("....Done")
    print()

for city in c :
    gradient_boosting("data/clean_twpoi_data/{}.csv".format(city), city)

def gradient_boosting(filepath, cityname):
    df = pd.read_csv(filepath)
    X = df.loc[:, df.columns != "predicted_accident"]
    y = df.loc[:, df.columns == "predicted_accident"]
    
    # Split in train/test
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.3, stratify = y, random_state = 42)
    
    

    # Standardisize
    sc_x = StandardScaler()
    X_train = sc_x.fit_transform(X_train)
    X_test = sc_x.transform(X_test)
    
    # Gradient Boosting Classifier
    print("{},  Gradient Boosting Classifier  ... ".format(cityname))
    parameters = {'n_estimators':[20, 30, 40, 50, 70, 100, 150, 200],
                  "learning_rate":[1.0, 0.9, 0.8, 0.7, 0.6, 0.5]}
    gb_clf = GradientBoostingClassifier()
    gb_clf_best = GridSearchCV(gb_clf, parameters)
    gb_clf_best.fit(X_train, y_train)
    print()
    print("Best parameters: ", gb_clf_best.best_params_)
    print()
    print("Accuracy score (training): {0:.3f}".format(gb_clf_best.best_estimator_.score(X_train, y_train)))
    print("Accuracy score (test): {0:.3f}".format(gb_clf_best.best_estimator_.score(X_test, y_test)))
    print()
    best_params = gb_clf_best.best_params_
    n_estimators = best_params['n_estimators']
    learning_rate = best_params['learning_rate']
    gb_clf2 = GradientBoostingClassifier(n_estimators = n_estimators, learning_rate=learning_rate, 
                                         random_state=0)
    gb_clf2.fit(X_train, y_train)
    predictions = gb_clf2.predict(X_test)

    print()
    print("Confusion Matrix:")
    print(confusion_matrix(y_test, predictions))
    
    print()
    print("Classification Report")
    print(classification_report(y_test, predictions))
    print("....Done")
    print()

for city in cities :
    gradient_boosting("data/clean_twpoi_data/{}.csv".format(city), city)  

#### 4-3- Artifical Neural Network (ANN)

def model_ann(filepath, cityname) :
    df = pd.read_csv(filepath)
    X = df.loc[:, df.columns != "predicted_accident"]
    y = df.loc[:, df.columns == "predicted_accident"]
    y = y.values
    # Split in train/test
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, stratify = y, random_state = 0)

    # Standardisize
    sc_x = StandardScaler()
    X_train = sc_x.fit_transform(X_train)
    X_test = sc_x.transform(X_test)
    # Modele ANN
    print("{},  Artifical Neural Network (ANN) train ... ".format(cityname))
    print()
    classifier = keras.Sequential([
                        keras.layers.Dense(512, activation = tf.nn.relu, input_dim = 30),
                        keras.layers.Dense(256, activation=tf.nn.relu),
                        #keras.layers.Dense(64, activation=tf.nn.relu),
                        keras.layers.Dense(1, activation=tf.nn.sigmoid),
                    ])
    # Compiling the ANN
    classifier.compile(optimizer='adam',
              loss='binary_crossentropy',
              metrics=['accuracy'])
    
    # Fitting the ANN to the Training set
    classifier.fit(X_train, y_train,  batch_size = 10, validation_split= 0.2,  epochs = 10)
    print("....Done")
    print()
    print("{},  Artifical Neural Network (ANN) Score test  ... ".format(cityname))
    classifier.evaluate(X_test, y_test)
    predictions = classifier.predict(X_test)
    predictions = (predictions>0.5)
    print("....Done")
    print()
    print("{},  Artifical Neural Network (ANN) Confusion Matrix et Classification Report  ... ".format(cityname))
    print()
    print("Confusion Matrix:")
    print(confusion_matrix(y_test, predictions))
    print()
    print("Classification Report")
    print(classification_report(y_test, predictions))
    print("....Done")
    print()


    

for city in cities :
    model_ann("data/clean_twpoi_data/{}.csv".format(city), city)
